# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is prototype of aquafarm management service",
        "title": "Aquafarm Management Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/farms": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get all farms with pagination and filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Farms"
                ],
                "summary": "Get all paginated farms",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Current Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Count of farms each page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "the sorts key of farms sortation",
                        "name": "sorts",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "the farms search by code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "the farms search by village",
                        "name": "village",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "the farms search by district",
                        "name": "district",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "the farms search by city",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "the farms search by province",
                        "name": "province",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "the farms search by postal code",
                        "name": "postalCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "time",
                        "description": "the farm created date is filtered greather than or equal created date start",
                        "name": "createdDateStart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "time",
                        "description": "the farm created date is filtered less than or equal created date end",
                        "name": "createdDateEnd",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.FarmPagingResponse"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create new farm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Farms"
                ],
                "summary": "Create new farm",
                "parameters": [
                    {
                        "description": "Farm Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFarmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.FarmResponse"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CreateFarmRequest": {
            "type": "object",
            "required": [
                "address",
                "city",
                "code",
                "description",
                "district",
                "loginData",
                "postalCode",
                "province",
                "village"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "loginData": {},
                "longitude": {
                    "type": "number"
                },
                "postalCode": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "village": {
                    "type": "string"
                }
            }
        },
...
        "dto.FarmPagingResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "currentPage": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FarmResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "sorts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "totalPage": {
                    "type": "integer"
                }
            }
        },
        "dto.FarmResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "ponds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PondResponse"
                    }
                },
                "postalCode": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "village": {
                    "type": "string"
                }
            }
        },
...
        "dto.PondResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "depth": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "farm": {
                    "$ref": "#/definitions/dto.FarmResponse"
                },
                "farmId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "long": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "wide": {
                    "type": "number"
                }
            }
        },
...
        "generateResp.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "generateResp.Success": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}